requirements.txt
fastapi==0.110.0
sqlalchemy==2.0.29 
uvicorn==0.29.0
python-dotenv==1.0.1
pytz==2024.1
psycopg2==2.9.9
cloudinary==1.39.1
pika==1.3.2
pillow==10.3.0
pydantic==2.6.4
python-dotenv==1.0.1
python-telegram-bot==21.1.1
requests==2.31.0